backend/python-api/pyproject.toml
[project]
name = "python-api"
version = "0.1.0"
description = "ShoppingList backend API based on FastAPI + PostgreSQL"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]

[project.scripts]
python-api = "app.main:app"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.7",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 140
target-version = "py312"
exclude = ["migrations"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
